// apps/frontend/src/pages/admin/models/admin-news-routing.enum.ts
export enum AdminNewsRouting {
    SIMPLE = 'simple',
    OFFLINE_START = 'offline-start',
    OFFLINE_RESULTS = 'offline-results',
    ONLINE_ANNOUNCE = 'online-announce',
    ONLINE_RESULTS = 'online-results',
    MULTICUP_RESULTS = 'multicup-results',
    DFWC_ROUND_RESULTS = 'dfwc-round-results',
    STREAMERS_RESULTS = 'streamers-results',
}